Update files to private git server version
-Use os.fork() so that both main.py and server.py don't have to be run separately. Now you have to only run main.py and the server should come up 10 seconds after forking.
-Renamed loads of variables in the code, tried to unify the meaning of every variable in the system.
-Changed a few object definitions in order to clarify the way they work
-Added more comments and print statements to the terminal
-Changed database structure to enable less frequent API calls
--Program will now check if another user has previously added the same code before into the database in the case of users sharing shipping codes. This is to avoid doing more API calls and to avoid issues with the code being added already into the service.
-Improved multi-user support
--Fixed an issue which allowed other users to delete each other tracking codes if they were named the same
-Added new commands:
--forceadd <discord_user_id> <tracking_code> <optional: carrier_code>
--forceremove <discord_user_id> <package_name>
--checkcode <package_name>
-Added disclaimer for new users
--Secondary database for registered users
-Added eula to the bot after first startup

+++ objects/database_object.py
@@ -1,6 +1,42 @@
 import sqlite3
 
 
+class DiscordUsers:
+    def __init__(self):
+        self.__cursor = None
+        self.__conn = None
+
+    def connect(self, name):
+        self.__conn = sqlite3.connect(name, timeout=10)
+        self.__cursor = self.__conn.cursor()
+
+    def init_database(self):
+        #   0 is false
+        #   1 is true
+        self.__cursor.execute("CREATE TABLE users (discord_user text, has_agreed integer)")
+
+    def add_user(self, user):
+        self.__cursor.execute("INSERT INTO users (discord_user, has_agreed) VALUES (?, ?)", (str(user), 1,))
+
+    def remove_user(self, user):
+        self.__cursor.execute("DELETE FROM users WHERE discord_user = ?", (str(user),))
+
+    def check_user(self, user):
+        self.__cursor.execute("SELECT * FROM users WHERE discord_user = ?", (str(user),))
+        user_in_db = self.__cursor.fetchone()
+
+        if user_in_db is None:
+            return False
+        else:
+            return True
+
+    def commit_changes(self):
+        self.__conn.commit()
+
+    def close(self):
+        self.__conn.close()
+
+
 class ShippingCodesDB:
     def __init__(self):
         self.__cursor = None
@@ -11,42 +47,38 @@ class ShippingCodesDB:
         self.__cursor = self.__conn.cursor()
 
     def init_database(self):
-        self.__cursor.execute("CREATE TABLE codes (package_owner text, delivery_code text, courier_code text, name text, last_update text, last_update_courier text)")
+        self.__cursor.execute("CREATE TABLE codes (package_owner text, delivery_code text, carrier_code text, name text, carrier_detect text)")
 
-    def add_shipping_code(self, package_owner, delivery_code, courier_code, name, last_update, last_update_courier):
-        self.__cursor.execute("INSERT INTO {tn} (package_owner, delivery_code, courier_code, name, last_update, last_update_courier) VALUES (?, ?, ?, ?, ?, ?)".format(tn="codes"),
-                              (str(package_owner), str(delivery_code), str(courier_code),
-                               str(name), str(last_update), str(last_update_courier)))
+    def add_shipping_code(self, package_owner, delivery_code, carrier_code, name, carrier_detect):
+        self.__cursor.execute("INSERT INTO codes (package_owner, delivery_code, carrier_code, name, carrier_detect) VALUES (?, ?, ?, ?, ?)",
+                              (str(package_owner), str(delivery_code), str(carrier_code),
+                               str(name), str(carrier_detect)))
 
-    def remove_shipping_code(self, name):
-        self.__cursor.execute('''DELETE FROM codes WHERE name = ?''', (name,))
+    def remove_shipping_code(self, user_id, name):
+        self.__cursor.execute('''DELETE FROM codes WHERE name = ? AND package_owner = ?''', (name, str(user_id)))
 
     def tracking_stop(self, code, carrier):
-        self.__cursor.execute('''DELETE FROM codes WHERE delivery_code = ? AND courier_code = ?''', (code, carrier,))
+        self.__cursor.execute('''DELETE FROM codes WHERE delivery_code = ? AND carrier_code = ?''', (code, carrier,))
 
-    def has_been_updated(self, name, timestamp):
-        self.__cursor.execute('''SELECT last_update FROM codes WHERE name = ?''', (name,))
-        last_update = self.__cursor.fetchone()
-        
-        if last_update[0] is not timestamp:
-            return True
-        return False
-
-    def has_courier_updated(self, name, timestamp):
-        self.__cursor.execute('''SELECT last_courier_update FROM codes WHERE name = ?''', (name,))
-        last_update = self.__cursor.fetchone()
-
-        if last_update[0] is not timestamp:
-            return True
-        return False
-
-    def get_name(self, code, carrier):
-        self.__cursor.execute('''SELECT name FROM codes WHERE delivery_code = (?) AND courier_code = (?)''', [str(code), str(carrier)])
-        name = self.__cursor.fetchone()
-        if name is not None:
-            name = name[0]
-
+    def get_names(self, code, carrier):
+        self.__cursor.execute('''SELECT name FROM codes WHERE delivery_code = (?) AND carrier_code = (?)''', [str(code), str(carrier)])
+        name = self.__cursor.fetchall()
+        print
         return name
+
+    def get_users(self, code, carrier):
+        self.__cursor.execute('''SELECT package_owner FROM codes WHERE delivery_code = (?) AND carrier_code = (?) ''', [str(code), str(carrier)])
+        users = self.__cursor.fetchall()
+
+        return users
+
+    def get_carrier_code(self, user_id, name):
+        self.__cursor.execute('''SELECT carrier_code FROM codes WHERE name = ? AND package_owner = ?''', (name, user_id,))
+        carrier_code = self.__cursor.fetchone()
+        if carrier_code is None:
+            return -1
+
+        return carrier_code[0]
 
     def get_code(self, name, user):
         self.__cursor.execute('''SELECT delivery_code FROM codes WHERE name = (?) AND package_owner = (?)''', [str(name), str(user)])
@@ -57,36 +89,20 @@ class ShippingCodesDB:
         return delivery_code[0]
 
     def get_all_codes(self, user):
-        self.__cursor.execute('''SELECT delivery_code, name FROM codes WHERE package_owner = (?)''', (user,))
+        self.__cursor.execute('''SELECT name FROM codes WHERE package_owner = ?''', (user,))
         codes = self.__cursor.fetchall()
 
         return codes
 
-    def get_user(self, code, carrier):
-        self.__cursor.execute('''SELECT package_owner FROM codes WHERE delivery_code = (?) AND courier_code = (?) ''', [str(code), str(carrier)])
-        users = self.__cursor.fetchone()
+    def get_carrier_data(self, tracking_code, carrier_code):
+        if carrier_code is None:
+            self.__cursor.execute('''SELECT carrier_code FROM codes WHERE delivery_code = ? AND carrier_detect = ?''',
+                                  (tracking_code, "automatic"))
+        else:
+            self.__cursor.execute('''SELECT carrier_code FROM codes WHERE delivery_code = ? AND carrier_code ''',
+                                  (tracking_code, carrier_code))
 
-        return users
-
-    def get_courier_code(self, name):
-        self.__cursor.execute('''SELECT courier_code FROM codes WHERE name = ?''', (name,))
-        courier_code = self.__cursor.fetchone()
-        if courier_code is None:
-            return -1
-
-        return courier_code[0]
-
-    def modify_last_update(self, name, last_update):
-        self.__cursor.execute('''UPDATE codes SET last_update = ? WHERE name = ?''',
-                              [str(last_update), str(name)])
-
-    def modify_last_courier_update(self, name, last_courier_update):
-        self.__cursor.execute('''UPDATE codes SET last_update_courier = ? WHERE name = ?''',
-                              [str(last_courier_update), str(name)])
-
-    def modify_courier_code(self, name, new_courier):
-        self.__cursor.execute('''UPDATE codes SET courier_code = ? WHERE name = ?''',
-                              [str(new_courier), str(name)])
+        return self.__cursor.fetchone()
 
     def commit_changes(self):
         self.__conn.commit()
